Create donkey.js
Just open your browser devtools, copy and paste this code. Press ENTER and follow the instructions.

1. Wait for 2 logs from firebase scripts to output and you can exec << useDonkey() >>
2. Select a target and start hack
3. If the asked word is in the database, your input will be filled automatically
4. You will just have to press ENTER to submit

If the asked word isn't in the database, juste type it in and Donkey will store it. So that the next time you won't have to type this same word again.

Donkey is participatory, let's work together, making the world a better place ;)